Updates
Updated Script to allow police (while in an emergency vehicle) to press G and have the ability to change a traffic light by putting the line drawn on it. There is also a new file (blacklist.lua) that contains whatever blacklisted models you do not want to be deleted upon Pool Check every 2 seconds.